+++ crates/nostr/src/util/mod.rs
@@ -22,7 +22,7 @@ pub mod hex;
 pub mod hkdf;
 
 use crate::nips::nip01::Coordinate;
-use crate::{EventId, PublicKey, SecretKey, Tag};
+use crate::{EventBuilder, EventId, PublicKey, SecretKey, Tag, UnsignedEvent};
 
 /// Generate shared key
 ///
@@ -117,5 +117,25 @@ impl From<EventId> for EventIdOrCoordinate {
 impl From<Coordinate> for EventIdOrCoordinate {
     fn from(coordinate: Coordinate) -> Self {
         Self::Coordinate(coordinate)
+    }
+}
+
+/// Event builder or unsigned event
+pub enum EventBuilderOrUnsignedEvent {
+    /// Event Builder
+    EventBuilder(EventBuilder),
+    /// Unsigned event
+    UnsignedEvent(UnsignedEvent),
+}
+
+impl From<EventBuilder> for EventBuilderOrUnsignedEvent {
+    fn from(value: EventBuilder) -> Self {
+        Self::EventBuilder(value)
+    }
+}
+
+impl From<UnsignedEvent> for EventBuilderOrUnsignedEvent {
+    fn from(value: UnsignedEvent) -> Self {
+        Self::UnsignedEvent(value)
     }
 }